Autonomous Seems to Be Working with Dashboard Values
All autonomous but the top goal with second drive added are working. The
second drive has not been tested yet. Teleop for the winch mechanism has
been added. Depending on which way it goes when we turn it we will then
negate the input. A second reset has been added for the encoder. Encoder
was changed from 1x to 4x because it was wrong and should have been 4x.